requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.14.0
asgiref==0.14.0
autobahn==0.15.0
billiard==3.3.0.23
celery==3.1.23
channels==0.17.0
daphne==0.14.2
Django==1.9.8
django-debug-toolbar==1.5
django-filter==1.0.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
kombu==3.0.35
msgpack-python==0.4.7
pytz==2016.6.1
redis==2.10.5
six==1.10.0
sqlparse==0.2.1
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.2.0
